SME-2 refactoring app.py and adjust UI of dialog
