+++ articledb/article.py
@@ -1,38 +1,47 @@
-def create_context(
-    question, df, max_len=1800, size="ada"
-):
+# import openai, os
+# import pandas as pd
+# import numpy as np
+# from ast import literal_eval
+# from articledb.article import answer_question
+from openai.embeddings_utils import distances_from_embeddings, cosine_similarity
+
+from articledb import openai, df
+
+def create_context(question, df, max_len=1800, size="ada"):
     """
     Create a context for a question by finding the most similar context from the dataframe
     """
 
     # Get the embeddings for the question
-    q_embeddings = openai.Embedding.create(input=question, engine='text-embedding-ada-002')['data'][0]['embedding']
+    q_embeddings = openai.Embedding.create(
+        input=question, engine='text-embedding-ada-002')['data'][0]['embedding']
 
     # Get the distances from the embeddings
-    df['distances'] = distances_from_embeddings(q_embeddings, df['embeddings'].values, distance_metric='cosine')
-
+    df['distances'] = distances_from_embeddings(
+        q_embeddings, df['embeddings'].values, distance_metric='cosine')
 
     returns = []
     cur_len = 0
 
     # Sort by distance and add the text to the context until the context is too long
     for i, row in df.sort_values('distances', ascending=True).iterrows():
-        
+
         # Add the length of the text to the current length
         cur_len += row['n_tokens'] + 4
-        
+
         # If the context is too long, break
         if cur_len > max_len:
             break
-        
+
         # Else add it to the text that is being returned
         returns.append(row["text"])
 
     # Return the context
     return "\n\n###\n\n".join(returns)
 
+
 def answer_question(
-    df,
+    df = df,
     model="text-davinci-003",
     question="Am I allowed to publish model outputs to Twitter, without a human review?",
     max_len=1800,
