+++ templates/home.html
@@ -162,7 +162,7 @@
         position: fixed;
         bottom: 10px;
         right: 10px;
-        height: 600px;
+        height: 650x;
         width: 400px;
         padding: 10px 0px;
         box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
@@ -178,8 +178,8 @@
       }
 
       #askSummerConversationHistoryDiv {
-        height: 480px;
-        max-height: 480px;
+        height: 530px;
+        max-height: 530px;
         overflow-y: auto;
         text-align: left;
       }
@@ -240,7 +240,7 @@
       <div id="askSummerConversationHistoryDiv">
       </div>
       <div id="askSummerQuestionDiv" class="input-group mb-3">
-        <textarea class="form-control" placeholder="Who is Summer?" id="askSummerTA"></textarea>
+        <textarea class="form-control" placeholder="Ask any question. A notification will be sent to Summer(Person) if AI unable to answer." id="askSummerTA" autofocus></textarea>
         <button class="btn btn-success" type="submit" onclick="answer()">Ask</button>
       </div>
     </div>
