+++ webex/__init__.py
@@ -1,0 +1,6 @@
+import os
+
+WEBEX_MSG_API_URL= 'https://webexapis.com/v1/messages'
+WEBEX_ACCESS_TOKEN = os.getenv("WEBEX_ACCESS_TOKEN")
+ASK_SUMMER_BOT_ID = os.getenv("ASK_SUMMER_BOT_ID")
+
